pyproject.toml
[project]
name = "productprice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.0",
    "Django==5.2.8",
    "djangorestframework==3.16.1",
    "et-xmlfile==2.0.0",
    "numpy==1.26.4",
    "openpyxl==3.1.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "python-dotenv==1.0.1",
    "psycopg2-binary==2.9.11",
    "polars[excel]==0.21.7",
]


